requirements.txt
PyOpenGL==3.1.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
